requirements.txt
streamlit
google-generativeai
python-dotenv
nltk
numpy
pandas
PyMuPDF
python-docx
